requirements.txt
requests == 2.7.0
python-docx == 0.8.10
pytest
